Allow additional sources to be specified as build arguments
Extract Vyatta specific sources.
Pass additional sources as a build argument. The benefit of this approach
over the old approach of moving the source.list file to the build context
(current folder) is that the user doesn't have to deal with that old
file lying about.
It may also make Jenkins integration better we won't have to check in
an additional file to the repo.

Things to note:
The default.sources uses the new deb822 format.
The embedded signed-by section is only available after apt 2.3.10 so
that's why I've had to bump the Debain version to 12.
The key has to have that dot (.) in it and it has to be indented by a
space on every line.
I wasted a lot of time not knowing those previous things...