Ubuntu requires docker-buildx package.
I added this to Recommends so the the installation doesn't fail on
debian as it won't be able to find the package (as it's ubuntu only).

There might be a better way to do this and maybe I should have two
separate .debs (an ubuntu one and a debian one). But for simplicity I'll
go with this solution for now.