Dockerfile: Split up copying to so each file is copied as a different stage
Previously there was a limitation with docker that when conditionally
copying then one of the files had to exist
(https://stackoverflow.com/a/46801962/13365272).

This no longer a limitation so each COPY can be done as separate stage.

The big benefit of this is it means docker can using the layering to
better cache the image. If the debian/control file is
changed then everything below that stage in the dockerfile must be rerun.
With these changes, I am able to push that COPY further down the dockerfile
so docker build has less stages to run.