Feature: Add --hook
Allow the user to specify a hook to run after the image is built.

The hooks will be stored in /etc/debpic/hooks

The user will specify the hook in a similar manner to how the sources
are specified.

Add initial gopath hook which allows devs to run "go test ./...", etc.

1. The python will copy the hook file (if it exists) to the build context.

2. The dockerfile will copy the (moved) hook to the image (if it exists)
```
COPY hoo[k] /usr/bin/hook
```

3. All modes (interactive, command specified & package commands) will
check if the hook exists and execute it if it does.
I'm currently executing the hook by using source so the bash exports
are maintained in the shell (for example export GOPATH=/usr/share/gocode/).
Unfortunately this means the scripts are limited to bash. This could be
revisited.
```
if [[ -x /usr/bin/hook ]]; then source /usr/bin/hook; fi
```