Volume mount: Only mount /home/docker/.cache
Mounting all of /home/docker was good as it meant bash history was
retained.
However when "debpic --vscode" was run it would create ~/.gnupg.
If ~/.gnupg exists then dpkg-buildpackage will try and sign the artifacts
and it will usually fail because the users email may not be one in the
changelog.

To ensure debpic "just builds" most repos I have stopped ~/.gnupg from
persisting. Of course some users may want to sign their artifacts so
this issue will need to be revisited.