Feature: Forward gpg socket from host to container
This allows dpkg-buildpackage to sign packages.
It also allows aptly to sign the generated apt repo.

Most linux distros have gpg installed already so we can assume it is
installed. In future it may be beneficial to check gpg on the host
machine first and then skip it if it is not installed.

The default build behaviour is now the same as dpkg-buildpackage.
This means if the user doesn't have the key for the email address in
the changelog then the signing stage will fail. This can be fixed but
adding `--no-sign`. eg `debpic -- --no-sign`.

This can be added to the config file so the user doesn't have to do
this all the time.